#HCMFLX-adicionado a escuta do rabbit ao remover registro, validado ticket e adicionado capicidade do room

+++ movie/src/main/java/com/movie/movie/entity/Movie.java
@@ -21,6 +21,6 @@ public class Movie {
     private String name;
 
     @Column(name = "classification")
-    private Long classification;
+    private Short classification;
 
 }
